+++ common/src/main/java/org/exodusstudio/stellaris/common/items/RoverItem.java
@@ -22,7 +22,7 @@ import net.minecraft.world.level.Level;
 import org.exodusstudio.stellaris.common.entities.vehicles.RoverEntity;
 import org.exodusstudio.stellaris.common.modules.Modules;
 import org.exodusstudio.stellaris.common.modules.rover.RoverModule;
-import org.exodusstudio.stellaris.common.modules.rover.RoverModules;
+import org.exodusstudio.stellaris.common.modules.rover.RoverUpgrades;
 import org.exodusstudio.stellaris.common.registries.DataComponentsRegistry;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -81,7 +81,7 @@ public class RoverItem extends Item implements FluidProvider.ITEM {
 
     @Override
     public @Nullable UniversalFluidItemStorage getFluidTank(@NotNull ItemStack stack) {
-        Modules<RoverModule> modules = stack.getOrDefault(DataComponentsRegistry.ROVER_MODULES.get(), RoverModules.empty());
+        Modules<RoverModule> modules = stack.getOrDefault(DataComponentsRegistry.ROVER_MODULES.get(), RoverUpgrades.empty()).roverModules();
         return new ItemFluidStorage(DataComponentsRegistry.FLUID_LIST.get(), stack, 1, RoverEntity.getTankCapacity(modules));
     }
 }