+++ common/src/main/java/org/exodusstudio/stellaris/common/items/modules/rocket/HydrogenFuelModuleItem.java
@@ -28,7 +28,13 @@ public class HydrogenFuelModuleItem extends Item implements RocketModule.CustomF
     }
 
     @Override
+    public boolean fitsRover() {
+        return true;
+    }
+
+    @Override
     public void appendHoverText(ItemStack stack, TooltipContext context, TooltipDisplay tooltipDisplay, Consumer<Component> tooltipAdder, TooltipFlag flag) {
         tooltipAdder.accept(Component.translatable("tooltip.item.stellaris.can_be_applied_to_rocket_module").withColor(Utils.getMinecraftColor("gray")));
+        tooltipAdder.accept(Component.translatable("tooltip.item.stellaris.can_be_applied_to_rover_module").withColor(Utils.getMinecraftColor("gray")));
     }
 }