+++ fabric/src/main/java/org/exodusstudio/stellaris/platform/fabric/RegistryPlatformImpl.java
@@ -1,6 +1,7 @@
 package org.exodusstudio.stellaris.platform.fabric;
 
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityDataRegistry;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.network.syncher.EntityDataSerializer;
 import net.minecraft.resources.Identifier;
 
@@ -8,5 +9,9 @@ public class RegistryPlatformImpl {
 
     public static <T> void registerEntityDataSerializer(Identifier location, EntityDataSerializer<T> serializer) {
         FabricEntityDataRegistry.register(location, serializer);
+    }
+
+    public static void addItemAlias(Identifier from, Identifier to) {
+        RegistryAliases.add(Registries.ITEM, from, to);
     }
 }