+++ changelog.md
@@ -3,4 +3,8 @@
 ## Changes
 - Improve entity rendering in the tablet
 - Display the lander model for the lander entry in the tablet
-- 
+- Clean some outdated data in the rover entry in the tablet
+
+## Removed
+- Rover cargo module (will be automatically replaced by the rocket equivalent)
+- Rover hydrogen motor module (will be automatically replaced by the rocket equivalent)